+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## v1.1.0 - 2026-03-10
+
+### Fixes
+- update on 2026.03.10 17:15:55
+- inserito sifiScanning CallBack da usare ad esempio con telegram per evitare conflitti
+- aggiunto il logPrefix string
+- renamed to lnWifiManager Release 1.0.2
+- aggiustamenti configurazione
+- preparazione README.md
+- ready to run versione studiata con gemini Release 1.0.1
+- ready to run Release 1.0.0
+- update D2026-02-24T20.49
+- update D2026-02-24T18.27
+- update D2026-02-24T18.24
+- update D2026-02-23T18.28
+- update D2026-02-23T17.11
+- apportati miglioramenti di stabilità e di print degli eventi e delle reti
+- prima delle modifiche di ottimizzazione
+- initial commit...
+- Initial commit
+
 ## v1.0.3 - 2026-03-10
 
 ### Fixes